refactor: funnel.types.ts 파일 관심사 분리
- 상수 정의 및 타입 추출(options.ts), 비즈니스 타입(steps.ts), 사용자 입력 검증(validation.ts), 메인 export(index.ts)
