+++ src/pages/generate/types/funnel/options.ts
@@ -66,48 +66,3 @@ export type ClosetType =
   (typeof MAIN_ACTIVITY_OPTIONS.CLOSET_TYPE)[keyof typeof MAIN_ACTIVITY_OPTIONS.CLOSET_TYPE]['value'];
 export type OtherFurnitures =
   (typeof MAIN_ACTIVITY_OPTIONS.OTHER_FURNITURES)[keyof typeof MAIN_ACTIVITY_OPTIONS.OTHER_FURNITURES]['value'];
-
-// 각 단계별 완성된 데이터 타입
-// Required<>: 모든 속성을 필수로 만드는 TS utility type
-export type CompleteHouseInfo = Required<ImageGenerateSteps['HouseInfo']>;
-export type CompleteHouseStructure = Required<
-  ImageGenerateSteps['HouseStructure']
->;
-export type CompleteInteriorTaste = Required<
-  ImageGenerateSteps['InteriorTaste']
->;
-// MainActivity(STEP4)의 '기타가구'는 필수 입력란이 아니므로 추후 수정 예정
-export type CompleteMainActivity = Required<ImageGenerateSteps['MainActivity']>;
-
-// Funnel Step 정의
-export type ImageGenerateSteps = {
-  HouseInfo: {
-    houseType?: HouseType;
-    roomType?: RoomType;
-    roomSize?: RoomSize;
-  };
-  HouseStructure: {
-    houseType: HouseType;
-    roomType: RoomType;
-    roomSize: RoomSize;
-    selectedHouseStructure?: number[];
-  };
-  InteriorTaste: {
-    houseType: HouseType;
-    roomType: RoomType;
-    roomSize: RoomSize;
-    selectedHouseStructure: number[];
-    selectedInteriorTaste?: number[];
-  };
-  MainActivity: {
-    houseType: HouseType;
-    roomType: RoomType;
-    roomSize: RoomSize;
-    selectedHouseStructure: number[];
-    selectedInteriorTaste: number[];
-    primaryUsage?: PrimaryUsage;
-    bedType?: BedType;
-    closetType?: ClosetType;
-    otherFurnitures?: OtherFurnitures[];
-  };
-};
